refactor: 機能とテストに分離した。
